+++ yandex/resource_yandex_mdb_kafka_cluster.go
@@ -445,12 +445,20 @@ func resourceYandexMDBKafkaClusterKafkaSettings() *schema.Resource {
 				Elem:     &schema.Schema{Type: schema.TypeString},
 				Set:      schema.HashString,
 				Optional: true,
-				Computed: true,
 			},
 			"offsets_retention_minutes": {
 				Type:         schema.TypeString,
 				ValidateFunc: ConvertableToInt(),
 				Optional:     true,
+			},
+			"sasl_enabled_mechanisms": {
+				Type: schema.TypeSet,
+				Elem: &schema.Schema{
+					Type:         schema.TypeString,
+					ValidateFunc: validateParsableValue(parseKafkaSaslMechanism),
+				},
+				Set:      schema.HashString,
+				Optional: true,
 			},
 		},
 	}
@@ -941,6 +949,7 @@ var mdbKafkaUpdateFieldsMap = map[string]string{
 	"config.0.kafka.0.kafka_config.0.replica_fetch_max_bytes":         "config_spec.kafka.kafka_config_{version}.replica_fetch_max_bytes",
 	"config.0.kafka.0.kafka_config.0.ssl_cipher_suites":               "config_spec.kafka.kafka_config_{version}.ssl_cipher_suites",
 	"config.0.kafka.0.kafka_config.0.offsets_retention_minutes":       "config_spec.kafka.kafka_config_{version}.offsets_retention_minutes",
+	"config.0.kafka.0.kafka_config.0.sasl_enabled_mechanisms":         "config_spec.kafka.kafka_config_{version}.sasl_enabled_mechanisms",
 	"config.0.unmanaged_topics":                                       "config_spec.unmanaged_topics",
 	"config.0.zookeeper.0.resources.0.resource_preset_id":             "config_spec.zookeeper.resources.resource_preset_id",
 	"config.0.zookeeper.0.resources.0.disk_type_id":                   "config_spec.zookeeper.resources.disk_type_id",