+++ CHANGELOG.md
@@ -10,6 +10,9 @@ FEATURES:
 * **New Data Source:** `yandex_compute_snapshot_schedule`
 * **New Resource:** `yandex_compute_snapshot_schedule`
 
+ENHANCEMENTS:
+* add `sasl_enabled_mechanisms` attribute in `yandex_mdb_kafka_cluster` resource and data source
+
 ## 0.81.0 (October 20, 2022)
 BUG FIXES:
 * compute: fix size validation in `resource_yandex_compute_disk`